Final version for formatting hopefully.

+++ cpp/lib/src/TextUtil.cc
@@ -348,7 +348,7 @@ inline bool IsWhiteSpace(const char ch) {
 
 inline std::string OctalEscape(const char ch) {
     char buf[1 + 3 + 1];
-    std::sprintf(buf, "\\%03o", static_cast<unsigned>(ch));
+    std::sprintf(buf, "\\%03o", ch);
     return buf;
 }
